requirements.txt
numpy==1.17.2
PyQt5==5.13.0
PyQt5-sip==4.19.18
PyYAML==5.1.1
requests==2.22.0
Pillow==6.0.0
pandas==0.24.2
PyBluez==0.22
object-detection==0.1
notebook==5.7.8
matplotlib==3.0.3
openslide-python==1.1.1
jsonpickle==1.2
labelImg==1.8.3
opencv-contrib-python==4.1.1.26
jupyter==1.0.0
jupyter-client==5.2.4
imgaug==0.3.0
jupyter-console==6.0.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
opencv-python==4.1.0.25
opencv-python-headless==4.1.1.26
scikit-image==0.15.0
scikit-learn==0.21.2
python-dateutil==2.8.0
pytz==2018.7
scipy==1.3.0
sklearn==0.0
tensorboard==1.14.0
protobuf==3.9.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
virtualenv==15.2.0